Перенос логики в контроллер, рефакторинг рендера дополнительных фильмов
